+++ Structure_Engine/Structure_Engine.csproj
@@ -118,11 +118,21 @@
     <Compile Include="Create\SectionProperties\AluminiumSection.cs" />
     <Compile Include="Create\Framing\ConstantFramingProperty.cs" />
     <Compile Include="Create\Elements\FEMesh.cs" />
+    <Compile Include="Create\SectionProperties\ConcreteSection\ConcreteCircularSection.cs" />
+    <Compile Include="Create\SectionProperties\ConcreteSection\ConcreteFreeFormSection.cs" />
+    <Compile Include="Create\SectionProperties\ConcreteSection\ConcreteRectangleSection.cs" />
+    <Compile Include="Create\SectionProperties\ConcreteSection\ConcreteSectionFromProfile.cs" />
+    <Compile Include="Create\SectionProperties\ConcreteSection\ConcreteTSection.cs" />
     <Compile Include="Create\SectionProperties\GenericSection.cs" />
     <Compile Include="Create\Requests\IResultRequest.cs" />
     <Compile Include="Create\MaterialFragments\Timber.cs" />
     <Compile Include="Create\SectionProperties\GeometricalSection.cs" />
-    <Compile Include="Create\SectionProperties\LongitudinalReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\BottomReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\LayerReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\LongitudinalReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\MultiLinearReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\PerimiterReinforcement.cs" />
+    <Compile Include="Create\SectionProperties\LongitudinalReinforcement\TopReinforcement.cs" />
     <Compile Include="Create\SectionProperties\TimberSection.cs" />
     <Compile Include="Modify\AddReinforcement.cs" />
     <Compile Include="Modify\SetMaterial.cs" />
@@ -170,7 +180,6 @@
     <Compile Include="Create\Elements\Panel.cs" />
     <Compile Include="Create\Elements\Node.cs" />
     <Compile Include="Create\Elements\Bar.cs" />
-    <Compile Include="Create\SectionProperties\ConcreteSection.cs" />
     <Compile Include="Create\Constraints\LinkConstraint.cs" />
     <Compile Include="Create\Elements\Opening.cs" />
     <Compile Include="Create\Elements\Surface.cs" />
